requirements.txt
pulumi>=3.133.0,<4
pulumi-datarobot>=0.2.5,<0.3

datarobot>=3.5.2
datarobot-predict>=1.9.2
datarobotx-idp>=0.2.11

kedro>=0.19
kedro-datasets>=2.1.0
kedro-viz

pydantic>=2.9.2
pydantic-settings>=2.5.2
eval_type_backport>=0.2.0
python-magic-bin==0.4.14

# Constrained by datarobot-drum
pandas>=2.0.3,<3

streamlit>=1.39.0,<2
joblib>=1.4.2,<2
openai>=1.47.1,<2

jupyterlab>=4.2.5,<5
papermill>=2.6.0,<3

babel>=2.16,<3

pytest==8.0.2
mypy==1.11.2
types-PyYAML==6.0.12.20240917
ruff==0.6.9

# All of this for a stupid custom rag model
datarobot-drum>=1.11.3
scikit-learn>=1.5.0
transformers>=4.42.4
langchain>=0.3.7
langchain-core>=0.3.15
langchain-community>=0.3.5
langchain-huggingface>=0.1.2
langchain-openai>=0.2.6
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
opencv-contrib-python-headless>=4.8.1.78
unstructured[all-docs]>=0.12.3
torch>=2.2.2

